+++ packages/gamut/src/ConnectedForm/ConnectedForm.tsx
@@ -1,5 +1,6 @@
 import React, { useEffect, useMemo } from 'react';
 import {
+  FieldValues,
   FormProvider,
   FormProviderProps,
   Mode,
@@ -94,13 +95,23 @@ export function ConnectedForm<Values extends FormValues<Values>>({
   watchedFields,
   ...rest
 }: ConnectedFormProps<Values>) {
-  // the below is fixed in react-hook-form v7: GM-466
-  // eslint-disable-next-line @typescript-eslint/unbound-method
-
-  const { handleSubmit, formState, reset, watch, ...methods } = useForm({
+  const {
+    clearErrors,
+    handleSubmit,
+    formState,
+    reset,
+    trigger,
+    watch,
+    ...methods
+  } = useForm<FieldValues>({
     defaultValues,
     mode: validation,
   });
+
+  const onError = async () => {
+    clearErrors();
+    await trigger();
+  };
 
   const isSubmitSuccessful = submitSuccessStatus(
     wasSubmitSuccessful,
@@ -138,13 +149,15 @@ export function ConnectedForm<Values extends FormValues<Values>>({
   return (
     <PropsProvider value={contextValues}>
       <FormProvider
+        clearErrors={clearErrors}
         handleSubmit={handleSubmit}
         formState={formState}
         reset={reset}
+        trigger={trigger}
         watch={watch}
         {...methods}
       >
-        <Form onSubmit={handleSubmit(onSubmit)} {...rest}>
+        <Form onSubmit={handleSubmit(onSubmit, onError)} {...rest}>
           {children}
         </Form>
       </FormProvider>