+++ packages/monaco/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.6.48](http://github.com///compare/@codecademy/monaco@0.6.47...@codecademy/monaco@0.6.48) (2022-06-07)
+
+**Note:** Version bump only for package @codecademy/monaco
+
+
+
+
+
+### [0.6.47](http://github.com///compare/@codecademy/monaco@0.6.46...@codecademy/monaco@0.6.47) (2022-06-06)
+
+**Note:** Version bump only for package @codecademy/monaco
+
+
+
+
+
 ### [0.6.46](http://github.com///compare/@codecademy/monaco@0.6.45...@codecademy/monaco@0.6.46) (2022-05-24)
 
 **Note:** Version bump only for package @codecademy/monaco